requirements-dev.txt
django-extensions==3.0.2
ipython==7.16.1
pytest-django==3.9.0
coverage==5.2
flake8==3.8.3

-r requirements.txt